requirements.txt
disnake
python-i18n
python-dotenv